modularize api-resources and cleanup
